Bug 1244049 - Part 1: Define scoped enum for CSSPseudoElement type. r=dbaron

+++ dom/animation/PseudoElementHashEntry.h
@@ -47,7 +47,9 @@ public:
     if (!aKey)
       return 0;
 
-    return mozilla::HashGeneric(aKey->mElement, aKey->mPseudoType);
+    // Convert the scoped enum into an integer while adding it to hash.
+    return mozilla::HashGeneric(aKey->mElement,
+                                static_cast<uint8_t>(aKey->mPseudoType));
   }
   enum { ALLOW_MEMMOVE = true };
